+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog (nionui-tool)
 0.3.22 (UNRELEASED)
 -------------------
 
+- Implement high quality image drawing in most cases.
+
+- Do not auto expand when using min width/height. Simplifies layout.
+
 - Improve handling of virtual environments on Linux.
 
 - Qt version 5.12 on Windows.